backend/requirements.txt
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
django==2.1.7
django-cors-headers==3.0.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.9.3
djangorestframework-jwt==1.11.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.0
openapi-codec==1.3.2
PyJWT==1.7.1
pytz==2019.1
requests==2.21.0
simplejson==3.16.0
sqlparse==0.3.0
uritemplate==3.0.0
psycopg2-binary==2.8.2
django-debug-toolbar==1.11

# latest compatible version of djoser (VULNERABILITY ISSUE - MEDIUM SEVERITY)
djoser==1.3.3

# latest compatible version of urllib3
urllib3==1.24.3
